💀 Replace scary characters
HSDP Security has identified a number of potential 
scary characters in custom JSON fields. This change
rids the custom field of these characters before
🚚. Not doing so will cause the entire LogEvent to
be rejected and lost.

For the benefit of the customer the scary characters
are replaced with `💀` so as to signal the 
original custom field was modified to pass
validation.