Here we are again, replacing more scary characters
For Fiesta / Kubernetes we are encountering (Container Image) version numbers with characters e.g. '@' which are not allowed. Filtering or replacing these characters at the fluent-bit level
is pretty hairy, so we opt to change this at the library level.
Currently, we are replacing all disallowed characters with '💀'.

The alternative of not doing this at the library level is to force every higher level logging system to deal with this (arbitrary) HSDP Logging limitation.

Note that practically every LogEvent field has a set of disallowed characters so to fully solve the issue we will have to add additional fields to the replaceMap and check each field in each message.